+++ FE/src/components/BookRecordsByDate/BookRecordsByDate.tsx
@@ -1,7 +1,11 @@
-import { DAYS } from "@/src/constants";
-import type { DailyBookRecord, bookRecord } from "@/src/types";
+import type { bookRecord } from "@/src/types";
+import { sortBookRecordsByDate, getFormalizedDate } from "@/src/utils";
 
-export default function DailyBookRecord({ bookRecords }: DailyBookRecord) {
+export default function DailyBookRecord({
+  bookRecords,
+}: {
+  bookRecords: bookRecord[];
+}) {
   const bookRecordsByDate = sortBookRecordsByDate(bookRecords);
 
   return (
@@ -35,42 +39,3 @@ function BookRecordItem({ bookRecord }: { bookRecord: bookRecord }) {
     </li>
   );
 }
-
-// 날짜별로 bookRecord를 묶어주는 로직.
-const sortBookRecordsByDate = (bookRecords: bookRecord[]) => {
-  const result: { date: string; bookRecords: bookRecord[] }[] = [];
-  bookRecords.forEach((bookRecord) => {
-    const { date } = bookRecord;
-
-    if (!result.length) {
-      return result.push({
-        date,
-        bookRecords: [bookRecord],
-      });
-    }
-
-    const last = result[result.length - 1];
-    if (last.date !== date) {
-      const newDate = {
-        date,
-        bookRecords: [bookRecord],
-      };
-      result.push(newDate);
-      return;
-    }
-
-    last.bookRecords.push(bookRecord);
-  });
-
-  return result;
-};
-
-const getFormalizedDate = (dateStr: string) => {
-  const dateInfo = new Date(dateStr);
-  const year = dateInfo.getFullYear();
-  const month = (dateInfo.getMonth() + 1).toString().padStart(2, "0");
-  const date = dateInfo.getDate();
-  const day = DAYS[dateInfo.getDay()];
-
-  return `${year}년 ${month}월 ${date}일 (${day})`;
-};
